==> assignment-02/code/q3/1_api.ipynb <==
import pandas as pd
def read_text_file(file_path,printflag=0):
    try:
        with open(file_path, 'r') as file:
            # Read the entire file contents into a string
            file_contents = file.read()
        if printflag:
            print("File contents:")
            print(file_contents)
        return file_contents
    except FileNotFoundError:
        return f"File not found: {file_path}"
    except Exception as e:
        return f"An error occurred: {str(e)}"

  


# stock data
import yfinance as yf
import os

def download_stock_data(company_name, start_date, end_date, data_dir):
    # Create a directory to store the data if it doesn't exist
    if not os.path.exists(data_dir):
        os.mkdir(data_dir)
    
    # Define the data file path
    data_file_path = f"{data_dir}/{company_name}_stock_data.csv"
    
    if os.path.isfile(data_file_path):
        print(f"Data for {company_name} already downloaded. Loading from {data_file_path}")
        # stock_data = yf.download(company_name, start=start_date, end=end_date)
        stock_data = pd.read_csv('../../data/stock_data/AAPL_stock_data.csv')
        return stock_data
    else:
        print(f"Downloading data for {company_name} from Yahoo Finance...")
        stock_data = yf.download(company_name, start=start_date, end=end_date)
        stock_data.to_csv(data_file_path)
        print(f"Data downloaded and saved to {data_file_path}")
        return stock_data

# Define the parameters
company_name = "AAPL"
start_date = "2022-10-31"
end_date = "2023-10-31"
data_directory = "../../data/stock_data"

# Call the function to download or load the data
stock_data = download_stock_data(company_name, start_date, end_date, data_directory)

# Now you have the data in the `aapl_data` DataFrame
print(stock_data)



# exploring the stock data
stock_data
type(stock_data) #pandas dataframe
stock_data.shape #(251,6)
# stock_data.head()
stock_data.describe()

stock_data.columns #['Open', 'High', 'Low', 'Close', 'Adj Close', 'Volume']
stock_data.index #['2022-10-31', ..., '2023-10-30']

# Accessing a specific row by label (using .loc)
# stock_data.loc['2023-10-30']

# Accessing a specific row by integer index (using .iloc)
stock_data.iloc[2]  # Access the third row (0-based index)

# --------------------------
# Accessing a Specific Column:

# Using square brackets
stock_data['Volume']

# Using dot notation
stock_data.Open
# ----------------------




# news data
# Usage example:
from newsapi import NewsApiClient
api_path = '../../data/newsapi_key.txt'
api_key= read_text_file(api_path,1).strip()
newsapi = NewsApiClient(api_key=api_key)



# print(api_key)
start_date = "2023-10-06"
end_date = "2023-10-30"

news_articles = newsapi.get_everything(
    q=company_name,
    from_param=start_date,
    to=end_date,
    language="en",
    sort_by="publishedAt",
    page_size=100
)

# Access the list of articles
articles = news_articles['articles']
news_data=pd.DataFrame(articles)

articles

# last_date
last_timestamp=news_data.iloc[-1].publishedAt
# last_timestamp.split('T').str[0]
last_date=last_timestamp.split('T')[0]
while last_date!=start_date:
    print("last_date is : ",last_date)
    end_date=last_date
    news_articles = newsapi.get_everything(
        q=company_name,
        from_param=start_date,
        to=end_date,
        language="en",
        sort_by="publishedAt",
        page_size=100
    )

    # Access the list of articles
    articles = news_articles['articles']
    news_data= news_data.append(pd.DataFrame(articles), ignore_index=True)
    last_timestamp=news_data.iloc[-1].publishedAt
    # last_timestamp.split('T').str[0]
    last_date=last_timestamp.split('T')[0]



# print(len(news_articles['articles']))
# print(articles)

type(news_data) #pandas dataframe
news_data.shape #(487,8)
news_data.head()
# news_data.describe()

# news_data.columns #['source', 'author', 'title', 'description', 'url', 'urlToImage', 'publishedAt', 'content']
# text1=news_data.iloc[0]

# news_data.publishedAt 
news_data.head()
# type(news_data)

# Remove duplicates based on specific columns
news_data = news_data.drop_duplicates('content')

# Reset the index to ensure it's consecutive
news_data.reset_index(drop=True, inplace=True)

news_data.shape

news_data.to_csv('../../data/news_data1.csv', index=False)  # Use your desired file path
